PiLiDAR-main/PiLiDAR-main/requirements.txt
numpy
matplotlib>=3.9.2
plotly>5.24.0
pyserial>=3.5
scipy>=1.12.0
open3d>=0.18.0  # or open3d-cpu
opencv-python>=4.10  # or opencv-headless-python
pye57>=0.4.12
exifread>=3.0.0
jupyter

quat>=1.0.0
mpu6050>=0.0.3

rpi-lgpio>=0.6
#rpi-hardware-pwm>=0.2.2